requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-SocketIO==5.3.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
python-dotenv==1.0.0
gunicorn==20.1.0
gevent==22.10.2
gevent-websocket==0.10.1
requests==2.31.0
PyJWT==2.8.0
pydub==0.25.1
aiohttp==3.8.5
websockets==11.0.3
python-engineio==4.5.1
python-socketio==5.8.0
eventlet==0.33.3
baidu-aip==2.2.18.0
openai==0.28.0  # 用于DeepSeek API调用
pymysql==1.1.0  # 如果需要MySQL支持
psycopg2-binary==2.9.7  # 如果需要PostgreSQL支持
redis==4.5.5  # 如果需要Redis支持
celery==5.3.0  # 如果需要异步任务队列
transformers==4.33.3
torch==2.0.1
ffmpeg-python==0.2.0
uWSGI==2.0.21
greenlet==2.0.2